+++ src/com/jpexs/decompiler/flash/gui/GenericTagTreePanel.java
@@ -45,6 +45,7 @@ import com.jpexs.decompiler.flash.types.RGB;
 import com.jpexs.decompiler.flash.types.RGBA;
 import com.jpexs.decompiler.flash.types.annotations.Conditional;
 import com.jpexs.decompiler.flash.types.annotations.ConditionalType;
+import com.jpexs.decompiler.flash.types.annotations.DottedIdentifier;
 import com.jpexs.decompiler.flash.types.annotations.EnumValue;
 import com.jpexs.decompiler.flash.types.annotations.EnumValues;
 import com.jpexs.decompiler.flash.types.annotations.HideInRawEdit;
@@ -57,6 +58,7 @@ import com.jpexs.decompiler.flash.types.annotations.UUID;
 import com.jpexs.decompiler.flash.types.annotations.parser.AnnotationParseException;
 import com.jpexs.decompiler.flash.types.annotations.parser.ConditionEvaluator;
 import com.jpexs.decompiler.flash.types.filters.CONVOLUTIONFILTER;
+import com.jpexs.decompiler.graph.DottedChain;
 import com.jpexs.helpers.ByteArrayRange;
 import com.jpexs.helpers.ConcreteClasses;
 import com.jpexs.helpers.ReflectionTools;
@@ -903,7 +905,10 @@ public class GenericTagTreePanel extends GenericTagPanel {
                     enumAdd = " - " + values.get(val);
                 }
 
-                if (val instanceof byte[]) {
+                DottedIdentifier di = field.getAnnotation(DottedIdentifier.class);
+                if (val instanceof String && di != null) {
+                    valStr += " = " + DottedChain.parseNoSuffix(val.toString()).toPrintableString(di.as3());
+                } else if (val instanceof byte[]) {
                     valStr += " = " + ((byte[]) val).length + " byte";
                 } else if (val instanceof ByteArrayRange) {
                     valStr += " = " + ((ByteArrayRange) val).getLength() + " byte";