+++ src/com/jpexs/decompiler/flash/gui/tagtree/TagTree.java
@@ -16,6 +16,7 @@
  */
 package com.jpexs.decompiler.flash.gui.tagtree;
 
+import com.jpexs.decompiler.flash.IdentifiersDeobfuscation;
 import com.jpexs.decompiler.flash.gui.AppStrings;
 import com.jpexs.decompiler.flash.gui.MainPanel;
 import com.jpexs.decompiler.flash.gui.TreeNodeType;
@@ -349,7 +350,7 @@ public class TagTree extends AbstractTagTree {
             String expName = tag.getSwf().getExportName(tag.getCharacterId());
             if (expName != null && !expName.isEmpty()) {
                 String[] pathParts = expName.contains(".") ? expName.split("\\.") : new String[]{expName};
-                return pathParts[pathParts.length - 1];
+                return IdentifiersDeobfuscation.printIdentifier(false, pathParts[pathParts.length - 1]);
             }
         }
         if (value != null) {